provider/vsphere: Linked cloning support.
Note: This is a breaking change!

The disk -> template part of the configuration is now a map instead of a string.

Linked cloning requires a snapshot to clone from. We support the ability to provide
a named snapshot. However, we try to use the current snapshot if no named snapshot is provided.

We error out if no snapshot name is provided and there is no current snapshot to clone from (error out if there is no snapshot available).
We error out if the provided snapshot name is not found.

The old template functionality is still available. You only need to specify the template label and then the old template functionality is executed (skip specifying the linked and snapshot values in the template configuration map).